backend/requirements.txt
amazon-braket-default-simulator==1.23.2
amazon-braket-schemas==1.21.4
amazon-braket-sdk==1.79.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.2
anyio==4.3.0
APScheduler==3.10.4
attrs==23.2.0
backoff==2.2.1
backports.entry-points-selectable==1.3.0
bcrypt==4.1.3
boltons==24.0.0
boto3==1.34.103
botocore==1.34.103
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
openpulse==0.5.0
openqasm3==0.5.0
opt-einsum==3.3.0
oqpy==0.3.5
orjson==3.10.3
outcome==1.3.0.post0
passlib==1.7.4
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
pymongo==4.7.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
s3transfer==0.10.1
scipy==1.13.0
selenium==4.20.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.37.2
sympy==1.12
trio==0.25.0
trio-websocket==0.11.1
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.29.0
watchfiles==0.21.0
websockets==12.0
wsproto==1.2.0
